refactor demos create a package structure to make it installable
