--- demos/cpu/cpu/systems/__init__.py
@@ -1,5 +1,0 @@
-from .cpu import CPU
-from .heatexchanger import HeatExchanger
-from .fancontroller import FanController
-from .fan import Fan
-from .cpusystem import CPUSystem
