--- demos/cpu/cpu/systems/cpu.py
@@ -1,24 +1,0 @@
-from cosapp.systems import System
-
-
-class CPU(System):
-    """ 
-    Evaluate thermal power generated by CPU
-    """
-    def setup(self):
-        # inputs
-        self.add_inward("usage", 20.0, desc="Usage percentage")
-        self.add_inward("tdp", 105.0, unit="W", desc="Thermal Design Power")
-        self.add_inward("heat_flow", 0.0, unit="W", desc="Exit thermal flow")
-        self.add_inward("heat_capacity", 20.0, "J/K", desc="CPU heat capacity")
-        self.add_inward("T", 20., unit="degC", desc="Metal temperature")
-
-        # outputs
-        self.add_outward("power", 20.0, unit="W", desc="Power")
-
-        # transients
-        self.add_transient("T", der="(power - heat_flow) / heat_capacity", desc="Enthalpy delta")
-
-
-    def compute(self):
-        self.power = self.tdp * self.usage / 100.0
