--- demos/cpu/cpu_lab.json
@@ -1,1 +1,0 @@
-{"modelJson": {"global": {"tabSetTabLocation": "bottom"}, "layout": {"type": "row", "id": "#1", "children": [{"type": "tabset", "id": "#2", "children": [{"type": "tab", "id": "#3", "name": "New section ", "component": "sub", "config": {"model": {"global": {}, "layout": {"type": "row", "id": "#1", "children": [{"type": "tabset", "id": "#11", "weight": 17.934119723424544, "children": [{"type": "tab", "id": "#4", "name": "Chart widget", "component": "grid", "config": {"layoutID": "#3"}}], "active": true}, {"type": "tabset", "id": "#6", "weight": 16.250164819816604, "children": [{"type": "tab", "id": "#5", "name": "Chart widget", "component": "grid", "config": {"layoutID": "#3"}}]}, {"type": "tabset", "id": "#9", "weight": 15.806543407787213, "children": [{"type": "tab", "id": "#8", "name": "Chart widget", "component": "grid", "config": {"layoutID": "#3"}}]}]}, "borders": []}}}], "active": true}]}, "borders": []}, "plotJson": {"#3@#4": {"traceConfig": {"time-cpu.usage:lines": {"name": "time-cpu.usage:lines", "plotType": "lines", "dataSource": "Recorders", "xAxisConfig": true, "systemNameX": "RK", "recorderRefX": "All", "xData": "time", "xDataIndex": "All", "yAxisConfig": true, "systemNameY": "RK", "recorderRefY": "All", "yData": "cpu.usage", "yDataIndex": "All", "zAxisConfig": true, "systemNameZ": "RK", "recorderRefZ": "All", "zData": "T_cpu", "zDataIndex": "All"}}, "plotLayout": {"title": {"text": "CPU usage"}, "xaxis": {"title": {"text": "time (s)"}, "exponentformat": "power", "autorange": true, "type": "linear", "range": [0, 30]}, "yaxis": {"title": {"text": "CPU usage (%)"}, "exponentformat": "power", "autorange": true, "type": "linear", "range": [-5.555555555555555, 105.55555555555556]}, "zaxis": {"title": {"text": "T_cpu"}, "exponentformat": "power", "autorange": true, "type": "normal", "range": ["", ""]}, "showlegend": true, "legend": {"orientation": "h"}, "autosize": true}, "advancedTraceDataConfig": {}}, "#3@#5": {"traceConfig": {"time-T_cpu:lines": {"name": "time-T_cpu:lines", "plotType": "lines", "dataSource": "Recorders", "xAxisConfig": true, "systemNameX": "RK", "recorderRefX": "All", "xData": "time", "xDataIndex": "All", "yAxisConfig": true, "systemNameY": "RK", "recorderRefY": "All", "yData": "T_cpu", "yDataIndex": "All", "zAxisConfig": true, "systemNameZ": "RK", "recorderRefZ": "All", "zData": "T_cpu", "zDataIndex": "All"}}, "plotLayout": {"title": {"text": "Temperature CPU"}, "xaxis": {"title": {"text": "time (s)"}, "exponentformat": "power", "autorange": true, "type": "linear", "range": [0, 30]}, "yaxis": {"title": {"text": "Temperature CPU (\u00b0C)"}, "exponentformat": "power", "autorange": true, "type": "linear", "range": [6.406534061227777, 78.27585283667223]}, "zaxis": {"title": {"text": "T_cpu"}, "exponentformat": "power", "autorange": true, "type": "normal", "range": ["", ""]}, "showlegend": true, "legend": {"orientation": "h"}, "autosize": true}, "advancedTraceDataConfig": {}}, "#3@#8": {"traceConfig": {"time-fan.tension:lines": {"name": "time-fan.tension:lines", "plotType": "lines", "dataSource": "Recorders", "xAxisConfig": true, "systemNameX": "RK", "recorderRefX": "All", "xData": "time", "xDataIndex": "All", "yAxisConfig": true, "systemNameY": "RK", "recorderRefY": "All", "yData": "fan.tension", "yDataIndex": "All", "zAxisConfig": true, "systemNameZ": "RK", "recorderRefZ": "All", "zData": "T_cpu", "zDataIndex": "All"}}, "plotLayout": {"title": {"text": "Tension Fan"}, "xaxis": {"title": {"text": "time (s)"}, "exponentformat": "power", "autorange": true, "type": "linear", "range": [0, 30]}, "yaxis": {"title": {"text": "Tension Fan (V)"}, "exponentformat": "power", "autorange": true, "type": "linear", "range": [-0.6666666666666667, 12.666666666666666]}, "zaxis": {"title": {"text": "T_cpu"}, "exponentformat": "power", "autorange": true, "type": "normal", "range": ["", ""]}, "showlegend": true, "legend": {"orientation": "h"}, "autosize": true}, "advancedTraceDataConfig": {}}}, "selectedVariable": {}}
