--- demos/cpu/parametric_geom_dashboard.json
@@ -1,1 +1,0 @@
-{"modelJson": {"global": {"tabSetTabLocation": "bottom"}, "layout": {"type": "row", "id": "#1", "children": [{"type": "tabset", "id": "#2", "children": [{"type": "tab", "id": "#3", "name": "New section ", "component": "sub", "config": {"model": {"global": {}, "layout": {"type": "row", "id": "#1", "children": [{"type": "tabset", "id": "#2", "children": [{"type": "tab", "id": "#4", "name": "3D widget", "component": "3Dview", "config": {"layoutID": "#3"}}], "active": true}]}, "borders": []}}}], "active": true}]}, "borders": []}, "plotJson": {}, "selectedVariable": {}}
