+++ requirements.txt
@@ -1,0 +1,4 @@
+cosapp
+plotly
+jupyterlab
+cosapp_lab
