+++ setup.cfg
@@ -1,0 +1,24 @@
+[metadata]
+name = twiinit-demos
+version = 0.0.1
+author = Adrien DELSALLE
+description = twiinIT demos based on simple use cases
+long_description = file: README.md, HISTORY.md
+home-page = http://github/twiinit/demos.git
+license = Proprietary License
+license-file = LICENSE
+keywords = demos, twiinit
+classifiers = 
+	Intended Audience :: End Users/Desktop
+	Intended Audience :: Developers
+	Intended Audience :: Science/Research
+	License :: Other/Proprietary License
+	Natural Language :: English
+	Programming Language :: Python :: 3
+
+[options]
+zip_safe = False
+include_package_data = True
+packages = find:
+python_requires = >= 3.8
+
