+++ twiinit_demos/cpu/systems/fan.py
@@ -1,0 +1,25 @@
+from cosapp.systems import System
+from twiinit_demos.cpu.ports.fluid import Fluid
+
+
+class Fan(System):
+    """
+    Evaluate air mass flow generated by fan
+    """
+
+    def setup(self):
+        # inputs
+        self.add_inward("T_air", 40.0, unit="degC", desc="Air temperature")
+        self.add_inward("tension", 0.0, unit="V", desc="Input tension")
+        self.add_inward("design_tension", 12.0, unit="V", desc="Fan input tension at design point")
+        self.add_inward("mass_flow_max", 1.0, unit="g/s", desc="Maximum mass flow")
+        self.add_inward("mass_flow_scalar", 1.0, unit="", desc="Percentage of mass flow available")
+
+        # outputs
+        self.add_output(Fluid, "fl_out")
+
+    def compute(self):
+        self.fl_out.mass_flow = (
+            self.mass_flow_scalar * self.mass_flow_max * self.tension / self.design_tension
+        )
+        self.fl_out.T = self.T_air
